requirements.txt
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
oauth2client==4.1.3
python-dotenv==1.0.0
pandas==2.2.1
matplotlib==3.8.4
numpy==1.26.4
seaborn==0.13.2
fpdf==1.7.2
ipywidgets==8.1.2
ipython==8.24.0
scikit-learn==1.5.0
scipy==1.13.0
streamlit==1.35.0
plotly==5.22.0
python-docx==1.1.2
python-pptx==0.6.23
xlsxwriter==3.2.0
reportlab==4.2.0
pillow==10.3.0